working on storable playlist order

+++ lib/Routes/Playlist/PlaylistUpdate.php
@@ -34,33 +34,35 @@ class PlaylistUpdate extends OpencastController
 
         $json = $this->getRequestData($request);
 
-        // delete all existing tags from the playlist
-        PlaylistTags::deleteByPlaylist_id($playlist->id);
+        if (isset($json['tags'])) {
+            // delete all existing tags from the playlist
+            PlaylistTags::deleteByPlaylist_id($playlist->id);
 
-        // readd the new ones
-        foreach ($json['tags'] as $new_tag) {
-            // check if tag already exists in oc_tags
+            // readd the new ones
+            foreach ($json['tags'] as $new_tag) {
+                // check if tag already exists in oc_tags
 
-            if ($new_tag['id']) {
-                $tag = Tags::find($new_tag['id']);
-            } else {
-                $tag = Tags::findOneBySQL('tag = ? AND user_id = ?', [$new_tag['tag'], $user->id]);
+                if ($new_tag['id']) {
+                    $tag = Tags::find($new_tag['id']);
+                } else {
+                    $tag = Tags::findOneBySQL('tag = ? AND user_id = ?', [$new_tag['tag'], $user->id]);
+                }
+
+                if (empty($tag)) {
+                    $tag = new Tags();
+                    $tag->tag     = $new_tag['tag'];
+                    $tag->user_id = $user->id;
+                    $tag->store();
+                }
+
+                $pltag = new PlaylistTags();
+                $pltag->playlist_id = $playlist->id;
+                $pltag->tag_id      = $tag->id;
+                $pltag->store();
             }
 
-            if (empty($tag)) {
-                $tag = new Tags();
-                $tag->tag     = $new_tag['tag'];
-                $tag->user_id = $user->id;
-                $tag->store();
-            }
-
-            $pltag = new PlaylistTags();
-            $pltag->playlist_id = $playlist->id;
-            $pltag->tag_id      = $tag->id;
-            $pltag->store();
+            unset($json['tags']);
         }
-
-        unset($json['tags']);
 
         $playlist->setData($json);
         $playlist->store();
